% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_lineage.R
\name{divide_lineage}
\alias{divide_lineage}
\title{Divide full pango lineage into individual parts}
\usage{
divide_lineage(pango_lineage_full)
}
\arguments{
\item{pango_lineage_full}{Full string of pango lineage without aliases.}
}
\value{
Tibble containing individual parts of the full pango lineage.
}
\description{
Divide full pango lineage into individual parts
}
\examples{
divide_lineage("B.1.1.529.5.3.1.1.1.1.1.1")
divide_lineage("B.1.617.2.2")
}
