% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alias_key.R
\name{read_alias_key}
\alias{read_alias_key}
\title{Read and prepare pango designation aliases}
\usage{
read_alias_key(
  link =
    "https://raw.githubusercontent.com/cov-lineages/pango-designation/master/pango_designation/alias_key.json",
  remove_recombinants = TRUE,
  remove_empty_translations = TRUE
)
}
\arguments{
\item{link}{Link to alias_key.json from the pango_designation repository.}

\item{remove_recombinants}{Logical, if recombinant aliases should be omitted.}

\item{remove_empty_translations}{Logical, if original haplotypes (empty translations) should be omitted.}
}
\value{
tibble containing columns pango_short and pango_long which can be used
to translate pango lineages.
}
\description{
Read and prepare pango designation aliases
}
\examples{
read_alias_key()
read_alias_key(remove_recombinants=FALSE)
read_alias_key(remove_empty_translations=FALSE)
}
