% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_lineage.R
\name{translate_lineage}
\alias{translate_lineage}
\title{Translate abbreviated pango lineage}
\usage{
translate_lineage(lineage)
}
\arguments{
\item{lineage}{Abbreviated pango lineage.}
}
\value{
Full sting of pango lineage without aliases.
}
\description{
Translate abbreviated pango lineage
}
\examples{
translate_lineage("BQ.1.1")
translate_lineage("AY.2")
}
