% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOC_color.R
\name{VOC_color}
\alias{VOC_color}
\title{Get font and background color based on VOC}
\usage{
VOC_color(pango_lineage_full)
}
\arguments{
\item{pango_lineage_full}{Full string of pango lineage without aliases.}
}
\value{
Named vector containing Hex code for font and background_level_base color.
}
\description{
Get font and background color based on VOC
}
\examples{
VOC_color("B.1.1.529.5.3.1.1.1.1.1.1")
VOC_color("B.1.617.2.2")
}
